+++ 3DRadSpace/Engine3DRadSpace/Graphics/Null/GraphicsDevice.hpp
@@ -20,7 +20,7 @@ namespace Engine3DRadSpace::Graphics::Null
 		std::unique_ptr<IVertexBuffer> _screenQuad;
 	public:
 		GraphicsDevice() = delete;
-		explicit GraphicsDevice(void* nativeWindowHandle, unsigned width = 800, unsigned height = 600);
+		explicit GraphicsDevice(void* nativeWindowHandle, size_t width = 800, size_t height = 600);
 
 		GraphicsDevice(GraphicsDevice&) = delete;
 		GraphicsDevice(GraphicsDevice&&) = delete;
